Update rules.json
VP amendment:

Old rule: The Vice President shall assume the duties of the President in their absence. The Vice President shall be the assistant chief supervisor of the BoD. Their duties include special objects and activities, community service activities, and member education. The Vice President shall also be the designated liaison between MemCom and the BoD.“